Add caching to affiliation search
